clean up dead code
