+++ Source/AwsCommonRuntimeKit/Utilities.swift
@@ -74,6 +74,14 @@ extension aws_byte_buf {
     }
 }
 
+extension aws_byte_cursor {
+    func toString() -> String? {
+        if self.len == 0 { return nil }
+        let data = Data(bytesNoCopy: self.ptr, count: self.len, deallocator: .none)
+        return String(data: data, encoding: .utf8)
+    }
+}
+
 extension aws_array_list {
     func byteCursorListToStringArray() -> [String] {
         var arrayList = self
@@ -113,63 +121,11 @@ public extension Int32 {
     }
 }
 
-//Todo: remove these pointers functions and use aws_allocator
-extension UnsafeMutablePointer {
-    func deinitializeAndDeallocate() {
-        self.deinitialize(count: 1)
-        self.deallocate()
-    }
-}
-
 extension Bool {
     var uintValue: UInt32 {
         return self ? 1 : 0
     }
 }
-
-//Todo: remove these pointers functions and use aws_allocator
-// Ensure that any UnsafeXXXPointer is ALWAYS initialized to either nil or a value in a single call. Prevents the
-// case where you create an UnsafeMutableWhatever and do not call `initialize()` on it, resulting in a non-null but
-// also invalid pointer
-func fromPointer<T, P: PointerConformance>(ptr: T) -> P {
-    let pointer = UnsafeMutablePointer<T>.allocate(capacity: 1)
-    pointer.initialize(to: ptr)
-    return P(OpaquePointer(pointer))
-}
-
-func fromOptionalPointer<T, P: PointerConformance>(ptr: T?) -> P? {
-    if let ptr = ptr {
-        return fromPointer(ptr: ptr)
-    }
-    return nil
-}
-
-func allocatePointer<T>(_ capacity: Int = 1) -> UnsafeMutablePointer<T> {
-    let ptr = UnsafeMutablePointer<T>.allocate(capacity: capacity)
-    zeroStruct(ptr)
-    return ptr
-}
-
-func toPointerArray<T, P: PointerConformance>(_ array: [T]) -> P {
-    let pointers = UnsafeMutablePointer<T>.allocate(capacity: array.count)
-
-    for index in 0..<array.count {
-        pointers.advanced(by: index).initialize(to: array[index])
-    }
-    return P(OpaquePointer(pointers))
-}
-
-protocol PointerConformance {
-    init(_ pointer: OpaquePointer)
-}
-
-extension UnsafeMutablePointer: PointerConformance {}
-
-extension UnsafeMutableRawPointer: PointerConformance {}
-
-extension UnsafePointer: PointerConformance {}
-
-extension UnsafeRawPointer: PointerConformance {}
 
 func withOptionalCString<Result>(
         to arg1: String?, _ body: (UnsafePointer<Int8>?) -> Result) -> Result {
