+++ Source/AwsCommonRuntimeKit/crt/ByteCursor.swift
@@ -3,10 +3,3 @@
 import AwsCCommon
 import Foundation
 
-extension aws_byte_cursor {
-    func toString() -> String? {
-        if self.len == 0 { return nil }
-        let data = Data(bytesNoCopy: self.ptr, count: self.len, deallocator: .none)
-        return String(data: data, encoding: .utf8)
-    }
-}
